Kickoff ai review

+++ bin/kick_off.sh
@@ -164,19 +164,27 @@ tests=0
 # Also chmod user.pem_test to be 500.
 #
 if [[ $ssh_key_file != "" ]]; then
-	cp $ssh_key_file config/user.pem_test
+	if [ ! -f $ssh_key_file ]; then
+		echo "${ssh_key_file} does not exists, please fix.  Test is exiting"
+		exit 1
+	fi
 	if [ ! -s $ssh_key_file ]; then
 		echo "${ssh_key_file} is zero length, please fix.  Test is exiting"
 		exit 1
 	fi
+	cp $ssh_key_file config/user.pem_test
 	chmod 500 config/user.pem_test
 fi
 if [[ $ssh_pub_key_file != "" ]]; then
-	cp $ssh_pub_key_file config/user.pub_test
+	if [ ! -f $ssh_pub_key_file ]; then
+		echo "${ssh_pub_key_file} does not exists, please fix.  Test is exiting"
+		exit 1
+	fi
 	if [ ! -s $ssh_pub_key_file ]; then
 		echo "${ssh_pub_key_file} is zero length, please fix.  Test is exiting"
 		exit 1
 	fi
+	cp $ssh_pub_key_file config/user.pub_test
 	chmod 400 config/user.pub_test
 fi
 
@@ -206,7 +214,7 @@ remove_perm_file()
 # Cycle through all the files setting system values.
 #
 export ANSIBLE_HOST_KEY_CHECKING=False
-echo "[defaults]" >> ansible.cfg
+echo "[defaults]" > ansible.cfg
 echo "roles_path = ~/.ansible/roles:/usr/share/ansible/roles:/etc/ansible/roles" >> ansible.cfg
 echo "log_path=${curdir}/ansible_log" >> ansible.cfg
 if [[ $ansible_noise_level != "normal" ]]; then
@@ -239,7 +247,7 @@ do
 	echo "  term_system: ${term_system}" >> ansible_vars.yml
 	system_type=`grep system_type  ansible_vars_main.yml | cut -d':' -f 2 | cut -d' ' -f 2`
 	attempts=1
-	while [ $attempts -ne $create_attempts ]
+	while [ $attempts -le $create_attempts ]
 	do
 		mkdir tf
 		echo ===== attempt $attempts of $create_attempts ==============
